+++ src/export/text.rs
@@ -1,5 +1,4 @@
-use std::fs::File;
-use std::io::{BufWriter, Write};
+use std::io::Write;
 
 use euphoxide::api::MessageId;
 use time::format_description::FormatItem;
@@ -14,16 +13,13 @@ const TIME_FORMAT: &[FormatItem<'_>] =
     format_description!("[year]-[month]-[day] [hour]:[minute]:[second]");
 const TIME_EMPTY: &str = "                   ";
 
-pub async fn export_to_file(
-    vault: &EuphRoomVault,
-    file: &mut BufWriter<File>,
-) -> anyhow::Result<()> {
+pub async fn export<W: Write>(vault: &EuphRoomVault, out: &mut W) -> anyhow::Result<()> {
     let mut exported_trees = 0;
     let mut exported_msgs = 0;
     let mut root_id = vault.first_root_id().await;
     while let Some(some_root_id) = root_id {
         let tree = vault.tree(some_root_id).await;
-        write_tree(file, &tree, some_root_id, 0)?;
+        write_tree(out, &tree, some_root_id, 0)?;
         root_id = vault.next_root_id(some_root_id).await;
 
         exported_trees += 1;
@@ -38,8 +34,8 @@ pub async fn export_to_file(
     Ok(())
 }
 
-fn write_tree(
-    file: &mut BufWriter<File>,
+fn write_tree<W: Write>(
+    out: &mut W,
     tree: &Tree<SmallMessage>,
     id: MessageId,
     indent: usize,
@@ -47,22 +43,22 @@ fn write_tree(
     let indent_string = "| ".repeat(indent);
 
     if let Some(msg) = tree.msg(&id) {
-        write_msg(file, &indent_string, msg)?;
+        write_msg(out, &indent_string, msg)?;
     } else {
-        write_placeholder(file, &indent_string)?;
+        write_placeholder(out, &indent_string)?;
     }
 
     if let Some(children) = tree.children(&id) {
         for child in children {
-            write_tree(file, tree, *child, indent + 1)?;
+            write_tree(out, tree, *child, indent + 1)?;
         }
     }
 
     Ok(())
 }
 
-fn write_msg(
-    file: &mut BufWriter<File>,
+fn write_msg<W: Write>(
+    file: &mut W,
     indent_string: &str,
     msg: &SmallMessage,
 ) -> anyhow::Result<()> {
@@ -85,7 +81,7 @@ fn write_msg(
     Ok(())
 }
 
-fn write_placeholder(file: &mut BufWriter<File>, indent_string: &str) -> anyhow::Result<()> {
+fn write_placeholder<W: Write>(file: &mut W, indent_string: &str) -> anyhow::Result<()> {
     writeln!(file, "{TIME_EMPTY} {indent_string}[...]")?;
     Ok(())
 }